rendering 2/999999

+++ src/rendering.c
@@ -43,7 +43,7 @@ int next_wrap(const char *s, int i, int goal) {
     return k;
 }
 
-void luia_render_text(SDL_Renderer *renderer, const char *text, rgba color, uint16_t x, uint16_t y, uint8_t size) {
+void luia_render_text(SDL_Renderer *renderer, const char *text, rgba color, int x, int y, uint8_t size) {
     SDL_Surface *surface = TTF_RenderText_Solid(fonts[size], text, (SDL_Color) {color.r, color.g, color.b, color.a});
     SDL_Texture *texture = SDL_CreateTextureFromSurface(renderer, surface);
 
@@ -58,7 +58,7 @@ void luia_render_text(SDL_Renderer *renderer, const char *text, rgba color, uint
     SDL_FreeSurface(surface);
 }
 
-void luia_render_box(SDL_Renderer *renderer, uint16_t x, uint16_t y, uint16_t w, uint16_t h, rgba color) {
+void luia_render_box(SDL_Renderer *renderer, int x, int y, int w, int h, rgba color) {
     SDL_Rect r = {x, y, w, h};
     SDL_SetRenderDrawColor(renderer, color.r, color.g, color.b, color.a);
     SDL_RenderFillRect(renderer, &r);
